Jenkinsfile: Enable CI for Maneki (fix multiple issues)
Fix1: media_tree git clone returning a 403

media_tree.git has been moved in the media subsystem refactoring.
Take this opportunity to use installed headers scripts instead.
Since we're running in docker we can't rely on 'uname -r', search for
the more recent kernel headers including checkpatch.pl.

Fix2: checkpatch failing if headers are updated in this run

$headers is evaluated before updating headers through apt, leading to
failure if they ever get updated.
Fix that and use the same query as KVERSION while at it.

Fix3: Only package debian branch as deb file

There are several reasons for doing so :
Jenkins Github plugin prevents the debianizer from accessing all the
repository branches, and so is unable to merge master with debian.
Previously master branch was packaged and uploaded to the unstable
channel of the apt repository. This was never used, remove this technical
debt and only package the debian branch for releases.
Moving to Github grants us automatic zip files for each tag, dropping
the need of the debianizer to zip.
Therefore move the package building directly to the Jenkinsfile instead
of using debianizer.

Signed-off-by: Sylvain Petinot <sylvain.petinot@foss.st.com>