dts: rpi: Rewrite device trees to support CM5
Previously, device tree overlays declared a reset-gpios property
responsible for handling sensor power on and power off. The reset-gpios
property defaulted to GPIO 5, which is historically the cam0-pwdn GPIO
for a lot of Raspberry Pi platforms, but not all of them.

For platforms not using GPIO 5, an override mechanism was in place.
For this we used of the '__overrides__' device tree node.

  target-path="/__overrides__";
    __overlay__ {
      cam0-pwdn-ctrl = <&wolfy>,"reset-gpios:0";
      cam0-pwdn      = <&wolfy>,"reset-gpios:4";
    };
  };

This instructs the firmware to replace the GPIO phandle in reset-gpios
by its cam0-pwdn-ctrl, and the GPIO number to cam0-pwdn.

As of CM5, cam0-pwdn-ctrl and cam0-pwdn parameter is no longer supplied
by the firmware. The correct way is now to use the cam1_reg node as a
regulator in the sensor node, the regulator will be powered on on
regulator_bulk_enable() by the driver and will effectively perform a
reset.

  vana-supply = <&cam1_reg>;

Regulator cam1_reg is the common regulator for all Raspberry Pis. For
rpi > 5 with 2 interfaces, we make use of device tree parameters to
target interface 0 instead if wanted :

  cam0 = <&i2c_frag>, "target:0=",<&i2c_csi_dsi0>,
         <&csi_frag>, "target:0=",<&csi0>,
         <&cam_node>, "clocks:0=",<&cam0_clk>,
         <&cam_node>, "vana-supply:0=",<&cam0_reg>;

Allowing the user to choose its interface directly in the config.txt
file with :

  dtoverlay=custom_overlay,camX

With this change we safely delete the split between '_cam0.dts' and
'_cam1.dts' overlays, and keep only one.

'rpi' folder contains some overlays that will be working for all rpi
platforms. 'rpi5' folder contains 4lanes overlays that will be working
for both rpi5 and cm5.

Signed-off-by: Benjamin Mugnier <benjamin.mugnier@foss.st.com>